:bug: Doesn't kick users when using the welcome cmd

+++ src/events/countrySelector/countrySelectorKick.ts
@@ -11,14 +11,15 @@ export default class CountrySelectorLeave extends Event {
     setInterval(async () => {
       const dataCursor = CountryModel.find({
         startedAt: { $lte: new Date(Date.now() + 24 * 60 * 60 * 1000) },
+        entry: true,
       }).cursor();
 
       dataCursor.on("data", async (selector: DocumentType<Country>) => {
         const channel = guild.channels.resolve(selector.channelId);
         const member = guild.members.resolve(selector.userId);
 
-        await channel.delete();
-        await member.kick();
+        if (channel) await channel.delete();
+        if (member.user.username === member.nickname) await member.kick();
         await CountryModel.deleteOne(selector);
       });
     }, 3 * 60 * 60 * 1000);
